requirements.txt
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9
requests>=2.32
trafilatura>=1.12